have crud for user

+++ src/main/java/com/jackdodev/newCo/user/UserRepository.java
@@ -2,4 +2,4 @@ package com.jackdodev.newCo.user;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Integer> {}
+public interface UserRepository extends CrudRepository<User, Long> {}
